[update]: refactored backward sync logic between pipeline stages using CondVar
Added new structure to the RiscSoc, called CommonDataBus (cdb) which holds a vector of mutex and condvar pairs.
Each pair is recommended to be used by an pipeline stage to store combinational data that other stages should react to.
One example is the branch logic, where once the address and condition of a branch instruction are computed in the EX stage, these values should be sent to the IF stage in order to change the current PC.
We are using the CondVar sync primitive in order to notifiy a stage that there is fresh hot data on the feedback wires which should be processed.